requirements.txt
torch
torchvision
hydra-core
lightning
numpy
omegaconf
wandb
scipy
